% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rbowtie_align_package_v2_20210424.R
\name{btAlign}
\alias{btAlign}
\title{Bowtie Aligner}
\usage{
btAlign(
  seqData = getwd(),
  orfLib = list.files(pattern = "*.fa$|*.fasta$", full.names = TRUE)[[1]],
  aligner = c("btAlign", "SpliceMap"),
  numCores = 1,
  keepFiles = FALSE,
  options = NULL
)
}
\arguments{
\item{seqData}{The  directory where Fastq files are located}

\item{orfLib}{The reference Fasta alignment file}

\item{numCores}{The number of cores to run on}

\item{keepFiles}{TRUE or FALSE, whether you want to keep the bowtie output files}

\item{options}{Extra options for bowtie that feed into "alignmentParameter"}
}
\value{
A list of aligned and unaligned counts dataframes, respectively
}
\description{
Generates counts using the original bowtie algorithm
}
\examples{
x <- btAlign(orfLib = fastaFileName,numCores = 6)
}
