% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rbowtie_align_package_v2_20210424.R
\name{btToCounts}
\alias{btToCounts}
\title{Convert Bowtie Output}
\usage{
btToCounts(df)
}
\arguments{
\item{df}{Output dataframe from an aligner in MIPSAlign}
}
\value{
Two dataframes in a list that contain the aligned and unaligned reads, respectively
}
\description{
Convert bowtie or bowtie2 output to counts
}
\examples{
newdf <- btToCounts(df)
}
