% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rbowtie_align_package_v2_20210424.R
\name{filterSeqs}
\alias{filterSeqs}
\title{Filter Fastq Files}
\usage{
filterSeqs(
  fileDir = getwd(),
  leftPattern = "",
  rightPattern = "",
  minLen = 20,
  trimTo = 40,
  numCores = 1
)
}
\arguments{
\item{fileDir}{The file directory with Fastq files, default is current working directory}

\item{leftPattern}{The 3' adapter to trim sequences by}

\item{rightPattern}{The 5' adapter to trim sequences by}

\item{minLen}{All sequences below this minimum length requirement are filtered out}

\item{trimTo}{Trims all reads to desired length (determined by the length of the longest sequence)}

\item{numCores}{The number of cores to run on}
}
\description{
Performs a variety of filtering steps to clean Fastq files
}
\examples{
filterSeqs(trimTo = 30)
}
