% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rbowtie_align_package_v2_20210424.R
\name{matchAlign}
\alias{matchAlign}
\title{Perfect Match Aligner}
\usage{
matchAlign(
  seqData = list.files(pattern = "*.fastq.gz$", full.names = TRUE),
  orfLib = list.files(pattern = "*.fa|.fasta", full.names = TRUE),
  numCores = 1
)
}
\arguments{
\item{seqData}{The list of Fastq files you want to align}

\item{orfLib}{The reference Fasta alignment file}

\item{numCores}{The number of cores to run on}
}
\value{
A list of aligned and unaligned counts dataframes, respectively
}
\description{
Generates counts on perfect sequence matches
}
\examples{
x <- matchAlign(orfLib = fastaFileName,numCores = 6)
}
