% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rbowtie_align_package_v2_20210424.R
\name{MIPSAlign}
\alias{MIPSAlign}
\title{MIPSAlign Trimmer and Aligner}
\usage{
MIPSAlign(
  fileDir = getwd(),
  refFile,
  filter = FALSE,
  leftPattern = "",
  rightPattern = "",
  minLen = 60,
  trimTo = 40,
  aligner = c("matchAlign", "btAlign", "bt2Align"),
  numCores = 1,
  keepFiles = FALSE,
  options = ""
)
}
\arguments{
\item{fileDir}{The file directory where all the Fastq files are located}

\item{refFile}{The reference Fasta alignment file}

\item{filter}{TRUE or FALSE, if you want to use the Fastq filter}

\item{leftPattern}{Relevant only if filter = TRUE, the 3' adapter to trim sequences by}

\item{rightPattern}{Relevant only if filter = TRUE, the 5' adapter to trim sequences by}

\item{minLen}{Relevant only if filter = TRUE, all sequences below this minimum length threshold are filtered out}

\item{trimTo}{Relevant only if filter = TRUE, trims all reads to desired length (determined by the length of the longest sequence)}

\item{aligner}{One of "matchAlign", "btAlign", or "bt2Align"}

\item{numCores}{The number of cores to run on}

\item{keepFiles}{TRUE or FALSE, whether you want to keep the bowtie output files}

\item{options}{Extra options for bowtie that feed into "alignmentParameter"}
}
\value{
A list of counts, unaligned counts, and possibly imperfect counts if a bowtie aligner was used
}
\description{
Combines trimming and aligning in one comprehensive function
}
\examples{
x <- MIPSAlign(refFile = fastaFileName,filter = TRUE,trimTo = 30,numCores = 6)
}
