% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rbowtie_align_package_v2_20210424.R
\name{ReadFasta}
\alias{ReadFasta}
\title{Read-in Fasta}
\usage{
ReadFasta(file)
}
\arguments{
\item{file}{The Fasta file location}
}
\value{
The dataframe with "name" and "sequence" columns
}
\description{
Allows you to read-in Fasta file as dataframe
}
\examples{
fa <- ReadFasta(getwd())
}
