% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rbowtie_align_package_v2_20210424.R
\name{WriteFasta}
\alias{WriteFasta}
\title{Write Fasta File}
\usage{
WriteFasta(data, filename)
}
\arguments{
\item{data}{The dataframe with "name" and "sequence" columns that you want stored}

\item{filename}{The name of the file output}
}
\value{
A Fasta file in appropriate format
}
\description{
Saves a dataframe as fasta file
}
\examples{
fa <- WriteFasta(getwd())
}
