% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rbowtie_align_package_v2_20210424.R
\name{bt2Align}
\alias{bt2Align}
\title{Bowtie2 Aligner}
\usage{
bt2Align(
  seqData = list.files(pattern = "*.fastq.gz$"),
  orfLib = list.files(pattern = "*.fa$|*.fasta$")[[1]],
  numCores = 1,
  keepFiles = FALSE,
  options = ""
)
}
\arguments{
\item{seqData}{The list of Fastq files you want to align}

\item{orfLib}{The reference Fasta alignment file}

\item{numCores}{The number of cores to run on}

\item{keepFiles}{TRUE or FALSE, whether you want to keep the bowtie output files}

\item{options}{Extra options that feed into bowtie2 "options"}
}
\value{
A list of aligned and unaligned counts dataframes, respectively
}
\description{
Generates counts using the bowtie2 algorithm
}
\examples{
x <- bt2Align(orfLib = fastaFileName,numCores = 6)
}
